add new controller and method AdminController transfer to new controller,and add unique=true for School name
