+++ src/Controller/AdminController.php
@@ -135,87 +135,4 @@ class AdminController extends AbstractController
             'student' => $user,
         ]);
     }
-
-    /**
-     * @Route("/nouvel_etablissement", name="new_school")
-     */
-    public function newSchool(Request $request): Response
-    {
-        $school = new School();
-        $form = $this->createForm(SchoolType::class, $school);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $school = $form->getData();
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($school);
-            $entityManager->flush();
-            $this->addFlash(
-                'success',
-                'L\'établissement a bien été créé!'
-            );
-            return $this->redirectToRoute('admin_new_school_level');
-        }
-        return $this->renderForm('admin/registration/newSchool.html.twig', [
-            'form' => $form,
-        ]);
-    }
-
-    /**
-     * @Route("/modification/{name}", name="edit_school")
-     */
-    public function editSchool(Request $request, School $school, EntityManagerInterface $entityManager): Response
-    {
-        $form = $this->createForm(SchoolType::class, $school);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
-            $this->addFlash(
-                'success',
-                'L\'établissement a bien été modifié!'
-            );
-
-            return $this->redirectToRoute('admin_series', [], Response::HTTP_SEE_OTHER);
-        }
-        return $this->renderForm('admin/registration/editSchool.html.twig', [
-            'school' => $school,
-             'form'  => $form,
-        ]);
-    }
-
-    /**
-     * @Route("/liste_des_etablissement", name="show_school")
-     */
-    public function showSchool(ManagerRegistry $managerRegistry): Response
-    {
-        $schools = $managerRegistry->getRepository(School::class)->findAll();
-
-        return $this->render('admin/registration/showSchool.html.twig', [
-            'schools' => $schools,
-        ]);
-    }
-
-    /**
-     * @Route("/nouvelle_classe", name="new_school_level")
-     */
-    public function newSchoolLevel(Request $request): Response
-    {
-        $schoolLevel = new SchoolLevel();
-        $form = $this->createForm(SchoolLevelType::class, $schoolLevel);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $schoolLevel = $form->getData();
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($schoolLevel);
-            $entityManager->flush();
-            $this->addFlash(
-                'success',
-                'La classe a bien été créé!'
-            );
-            return $this->redirectToRoute('admin_series');
-        }
-        return $this->renderForm('admin/registration/newSchoolLevel.html.twig', [
-            'form' => $form,
-            ]);
-    }
 }
