+++ templates/admin/admin_component/_nav.html.twig
@@ -22,17 +22,17 @@
                     <a class="nav-link">Créer un compte enseignant</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link  {{ app.request.pathInfo == path('word_new') ? 'active' : '' }}" href="{{ path('admin_select_school') }}">Créer un compte élève</a>
+                    <a class="nav-link  {{ app.request.pathInfo == path('word_new') ? 'active' : '' }}" href="{{ path('admin_registration_select_school') }}">Créer un compte élève</a>
                 </li>
                 <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle {{ app.request.pathInfo == path('admin_show_school') ? 'active' : '' }}" href="#" id="navbarScrollingDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    <a class="nav-link dropdown-toggle {{ app.request.pathInfo == path('admin_school_show_school') ? 'active' : '' }}" href="#" id="navbarScrollingDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Gestion des établissements
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarScrollingDropdown">
-                        <li><a class="dropdown-item" href="{{ path('admin_new_school') }}">Création d'une école</a></li>
-                        <li><a class="dropdown-item" href="{{ path('admin_show_school') }}">Modification d'une école</a></li>
+                        <li><a class="dropdown-item" href="{{ path('admin_school_new_school') }}">Création d'une école</a></li>
+                        <li><a class="dropdown-item" href="{{ path('admin_school_show_school') }}">Modification d'une école</a></li>
                         <li><hr class="dropdown-divider"></li>
-                        <li><a class="dropdown-item" href="{{ path('admin_new_school_level') }}">Création d'une classe</a></li>
+                        <li><a class="dropdown-item" href="{{ path('admin_school_level_new') }}">Création d'une classe</a></li>
                     </ul>
                 </li>
             </ul>
